Misc bug fixes: single-instance file lock, retry caps, range parsing
- MainPanel.checkAppIsRunning: single-instance detection now uses a
  per-user file lock (FileChannel.tryLock on
  ~/.megabasterd<VER>/.megabasterd.lock) instead of "is port 1338
  bound?". Foreign processes that happen to hold 1338 no longer
  cause MegaBasterd to silently exit thinking another instance is
  running (GH #717). Watchdog listener now scans WATCHDOG_PORT..PORT+9
  so a busy port doesn't break "bring existing instance to front"
  either. Also fixed an existing socket-leak in the watchdog accept
  loop and made the loop exit cleanly when the listener closes.
- MiscTools.formatBytes(Long): null-guard the autoboxing path. Was
  NPE'ing in the GUI thread on any provision-error code path that
  passed a not-yet-initialised _file_size, leaving the UI status as
  a cryptic "I/O ERROR null".
- Download.calculateLastWrittenChunk / Upload.calculateLastUploadedChunk:
  use (double) instead of (float) for the bytes->chunk-id divide.
  float has 24-bit mantissa; resumes of >16 MiB files lost precision
  on the boundary calculation and could rewind a chunk's worth of
  data, triggering MAC verify failure on otherwise good downloads.
- KissVideoStreamServer._parseRangeHeader: support RFC 7233 suffix
  ranges ("bytes=-N") and open-ended ("bytes=N-"). HTML5 video
  players issue suffix ranges when seeking near the end of a file;
  the old regex rejected them and the server fell through to
  streaming from offset 0 (broken seek).
- KissVideoStreamServer _getMegaFileMetadata / getMegaFileDownloadUrl:
  cap APIException retries at MAX_LINK_RETRIES (20) and propagate
  the FATAL_API_ERROR_CODES set (deleted, taken down, blocked, ...)
  immediately instead of looping at ~8s intervals forever. A user
  pasting a dead stream link no longer hangs the streamer worker.
  Restore interrupt status on InterruptedException.
- MegaAPI.RAW_REQUEST: cap retries at MAX_RAW_REQUEST_RETRIES (30).
  Persistent HTTP 500 from MEGA or unending -3/-4 used to spin
  forever at 8s intervals with no UI feedback. Logs the give-up
  decision.

+++ src/main/java/com/tonikelope/megabasterd/Download.java
@@ -238,7 +238,7 @@ public class Download implements Transference, Runnable, SecureSingleThreadNotif
 
     public long calculateLastWrittenChunk(long temp_file_size) {
         if (temp_file_size > 3584 * 1024) {
-            return 7 + (long) Math.floor((float) (temp_file_size - 3584 * 1024) / (1024 * 1024 * (this.isUse_slots() ? Download.CHUNK_SIZE_MULTI : 1)));
+            return 7 + (long) Math.floor((double) (temp_file_size - 3584 * 1024) / (1024 * 1024 * (this.isUse_slots() ? Download.CHUNK_SIZE_MULTI : 1)));
         } else {
             long i = 0, tot = 0;
